- Addressing review note. Finished OOP design for main script.
